refactor: 포스트 기본정보 분리
