+++ src/components/CreatePost/index.ts
@@ -1,0 +1,3 @@
+export { default as ComplexButton } from './ComplexButton';
+export { default as DatePicker } from './DatePicker';
+export { default as Location } from './Location';
