+++ src/components/common/index.ts
@@ -1,0 +1,2 @@
+export { default as SubTitle } from './SubTitle';
+export { default as Title } from './Titlte';
